chore: fix iamAuthOptions type
