+++ apps/studio/src/common/appdb/models/base.ts
@@ -1,9 +1,16 @@
-
-
 export interface QueryLike {
   title?: string
   text: string
   database?: string | null
   status?: string
+}
 
+export interface IamAuthOptions {
+  iamAuthenticationEnabled?: boolean
+  accessKeyId?: string;
+  secretAccessKey?: string;
+  awsRegion?: string;
+  clusterIdentifier?: string;
+  databaseGroup?: string;
+  tokenDurationSeconds?: number;
 }
